pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "industrial-3d-recon"
version = "0.1.0"
description = "CPU-only semi-automatic 3D reconstruction portfolio project for industrial parts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "imageio>=2.33",
  "imageio-ffmpeg>=0.4.9",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "open3d>=0.18",
  "PyYAML>=6.0",
  "scipy>=1.11",
  "trimesh>=4.0",
]

[project.optional-dependencies]
render3d = ["pyvista>=0.43"]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["recon*"]
